Module Tjr_lib__Tjr_config

A simple config file functor.

By default, we use the value of the env var "TJR_CONFIGS_DIR" as the location of config files. If this is not set, we use "/tmp".

val configs_dir : unit -> string

module Make : functor (S : sig ... end) -> sig ... end

Given a config type, and a filename, bind the value config to the Yojson-deserialized contents of the file.